+++ application/api/src/main/kotlin/io/raemian/api/config/JwtSecurityConfig.kt
@@ -1,0 +1,19 @@
+package io.raemian.api.config
+
+import io.raemian.api.support.JwtFilter
+import io.raemian.api.support.TokenProvider
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.web.DefaultSecurityFilterChain
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+
+class JwtSecurityConfig(
+    private val tokenProvider: TokenProvider,
+) : SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity>() {
+
+    // TokenProvider 를 주입받아서 JwtFilter 를 통해 Security 로직에 필터를 등록
+    override fun configure(http: HttpSecurity) {
+        val customFilter = JwtFilter(tokenProvider)
+        http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter::class.java)
+    }
+}
