+++ application/api/src/main/kotlin/io/raemian/api/controller/auth/request/SignInRequest.kt
@@ -1,0 +1,6 @@
+package io.raemian.api.controller.auth.request
+
+data class SignInRequest(
+    val email: String,
+    val password: String,
+)
