+++ application/api/src/main/kotlin/io/raemian/api/controller/auth/request/SignUpRequest.kt
@@ -1,0 +1,6 @@
+package io.raemian.api.controller.auth.request
+
+data class SignUpRequest(
+    val email: String,
+    val password: String,
+)
