+++ application/api/src/main/kotlin/io/raemian/api/domain/auth/OAuthProvider.kt
@@ -1,0 +1,5 @@
+package io.raemian.api.domain.auth
+
+enum class OAuthProvider {
+    GOOGLE, NAVER
+}
