+++ application/api/src/main/kotlin/io/raemian/api/domain/auth/TokenDTO.kt
@@ -1,0 +1,8 @@
+package io.raemian.api.domain.auth
+
+data class TokenDTO(
+    val grantType: String,
+    val accessToken: String,
+    val refreshToken: String,
+    val accessTokenExpiresIn: Long,
+)
