+++ application/api/src/main/kotlin/io/raemian/api/support/JwtFilter.kt
@@ -1,0 +1,43 @@
+package io.raemian.api.support
+
+import jakarta.servlet.FilterChain
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.util.StringUtils
+import org.springframework.web.filter.OncePerRequestFilter
+
+class JwtFilter(
+    private val tokenProvider: TokenProvider,
+) : OncePerRequestFilter() {
+
+    private val AUTHORIZATION_HEADER = "Authorization"
+    private val BEARER_PREFIX = "Bearer "
+    override fun doFilterInternal(
+        request: HttpServletRequest,
+        response: HttpServletResponse,
+        filterChain: FilterChain,
+    ) {
+        // 1. Request Header 에서 토큰을 꺼냄
+        val jwt: String = resolveToken(request)
+
+        // 2. validateToken 으로 토큰 유효성 검사
+        // 정상 토큰이면 해당 토큰으로 Authentication 을 가져와서 SecurityContext 에 저장
+        if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
+            val authentication: Authentication = tokenProvider.getAuthentication(jwt)
+            SecurityContextHolder.getContext().authentication = authentication
+        }
+        filterChain.doFilter(request, response)
+    }
+
+    // Request Header 에서 토큰 정보를 꺼내오기
+    private fun resolveToken(request: HttpServletRequest): String {
+        val bearerToken = request.getHeader(AUTHORIZATION_HEADER)
+        return if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
+            bearerToken.split(" ".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()[1].trim { it <= ' ' }
+        } else {
+            ""
+        }
+    }
+}
