+++ application/api/src/main/kotlin/io/raemian/api/support/error/ErrorCode.kt
@@ -1,0 +1,5 @@
+package io.raemian.api.support.error
+
+enum class ErrorCode {
+    E500,
+}
