+++ application/api/src/main/kotlin/io/raemian/api/support/response/ResultType.kt
@@ -1,0 +1,5 @@
+package io.raemian.api.support.response
+
+enum class ResultType {
+    SUCCESS, ERROR
+}
