--- application/api/src/main/kotlin/io/raemian/core/api/controller/v1/request/ExampleRequestDto.kt
@@ -1,11 +1,0 @@
-package io.raemian.core.api.controller.v1.request
-
-import io.raemian.core.api.domain.ExampleData
-
-data class ExampleRequestDto(
-    val data: String,
-) {
-    fun toExampleData(): ExampleData {
-        return ExampleData(data, data)
-    }
-}
