--- application/api/src/main/kotlin/io/raemian/core/api/domain/ExampleData.kt
@@ -1,6 +1,0 @@
-package io.raemian.core.api.domain
-
-data class ExampleData(
-    val value: String,
-    val param: String,
-)
