--- application/api/src/main/kotlin/io/raemian/core/api/domain/ExampleResult.kt
@@ -1,5 +1,0 @@
-package io.raemian.core.api.domain
-
-data class ExampleResult(
-    val data: String,
-)
