--- application/api/src/main/kotlin/io/raemian/core/api/domain/ExampleService.kt
@@ -1,10 +1,0 @@
-package io.raemian.core.api.domain
-
-import org.springframework.stereotype.Service
-
-@Service
-class ExampleService() {
-    fun processExample(exampleData: ExampleData): ExampleResult {
-        return ExampleResult(exampleData.value)
-    }
-}
