--- application/api/src/main/kotlin/io/raemian/core/api/support/error/CoreApiException.kt
@@ -1,6 +1,0 @@
-package io.raemian.core.api.support.error
-
-class CoreApiException(
-    val errorType: ErrorType,
-    val data: Any? = null,
-) : RuntimeException(errorType.message)
