--- application/api/src/main/kotlin/io/raemian/core/api/support/error/ErrorCode.kt
@@ -1,5 +1,0 @@
-package io.raemian.core.api.support.error
-
-enum class ErrorCode {
-    E500,
-}
