--- application/api/src/main/kotlin/io/raemian/core/api/support/response/ApiResponse.kt
@@ -1,24 +1,0 @@
-package io.raemian.core.api.support.response
-
-import io.raemian.core.api.support.error.ErrorMessage
-import io.raemian.core.api.support.error.ErrorType
-
-data class ApiResponse<T> private constructor(
-    val result: ResultType,
-    val data: T? = null,
-    val error: ErrorMessage? = null,
-) {
-    companion object {
-        fun success(): ApiResponse<Any> {
-            return ApiResponse(ResultType.SUCCESS, null, null)
-        }
-
-        fun <S> success(data: S): ApiResponse<S> {
-            return ApiResponse(ResultType.SUCCESS, data, null)
-        }
-
-        fun <S> error(error: ErrorType, errorData: Any? = null): ApiResponse<S> {
-            return ApiResponse(ResultType.ERROR, null, ErrorMessage(error, errorData))
-        }
-    }
-}
