--- application/api/src/main/kotlin/io/raemian/core/api/support/response/ResultType.kt
@@ -1,5 +1,0 @@
-package io.raemian.core.api.support.response
-
-enum class ResultType {
-    SUCCESS, ERROR
-}
