--- application/api/src/main/kotlin/io/raemian/core/auth/config/CorsConfig.kt
@@ -1,22 +1,0 @@
-package io.raemian.core.auth.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.cors.CorsConfiguration
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource
-import org.springframework.web.filter.CorsFilter
-
-@Configuration
-class CorsConfig {
-    @Bean
-    fun corsFilter(): CorsFilter {
-        val source = UrlBasedCorsConfigurationSource()
-        val config = CorsConfiguration()
-        config.allowCredentials = true
-        config.allowedOriginPatterns = listOf("*")
-        config.allowedHeaders = listOf("*")
-        config.allowedMethods = listOf("*")
-        source.registerCorsConfiguration("/**", config)
-        return CorsFilter(source)
-    }
-}
