--- application/api/src/main/kotlin/io/raemian/core/auth/config/JwtSecurityConfig.kt
@@ -1,19 +1,0 @@
-package io.raemian.core.auth.config
-
-import io.raemian.core.auth.support.JwtFilter
-import io.raemian.core.auth.support.TokenProvider
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.web.DefaultSecurityFilterChain
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-
-class JwtSecurityConfig(
-    private val tokenProvider: TokenProvider,
-) : SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity>() {
-
-    // TokenProvider 를 주입받아서 JwtFilter 를 통해 Security 로직에 필터를 등록
-    override fun configure(http: HttpSecurity) {
-        val customFilter = JwtFilter(tokenProvider)
-        http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter::class.java)
-    }
-}
