--- application/api/src/main/kotlin/io/raemian/core/auth/controller/v1/request/SignUpRequest.kt
@@ -1,6 +1,0 @@
-package io.raemian.core.auth.controller.v1.request
-
-data class SignUpRequest(
-    val email: String,
-    val password: String,
-)
