--- application/api/src/main/kotlin/io/raemian/core/auth/domain/OAuthProvider.kt
@@ -1,5 +1,0 @@
-package io.raemian.core.auth.domain
-
-enum class OAuthProvider {
-    GOOGLE, NAVER
-}
