--- application/api/src/main/kotlin/io/raemian/core/auth/domain/TokenDTO.kt
@@ -1,8 +1,0 @@
-package io.raemian.core.auth.domain
-
-data class TokenDTO(
-    val grantType: String,
-    val accessToken: String,
-    val refreshToken: String,
-    val accessTokenExpiresIn: Long,
-)
